refactor: rename Signin directory to Login and fix file name case issues
